media-libs/libva: Fix circular dependency with mesa
This removes the GLX backend to drop the "virtual/opengl" dependency.
Without removing this, it would pull in mesa which in turn would pull in
libva if vaapi support was turned on.

Removing the GLX backend doesn't seem to have any practical downsides,
even under X11, as the EGL backend seems to be used even if libva were
compiled with GLX support.

Signed-off-by: Sebastian Parborg <darkdefende@gmail.com>
Signed-off-by: Matt Turner <mattst88@gentoo.org>