net-firewall/nftables: Don't test iptables-nft rulesets in pkg_preinst()
Rulesets generated by iptables-nft are special in nature and will not
always be printed in a way that constitutes a valid syntax for nft(8).
Consider the following example in which iptables-nft would ideally have
generated a native rule that specifies "reject with tcp reset". Instead,
it generated a rule that integrates with an xtables target.

# iptables-nft -S
-A INPUT -j REJECT --reject-with tcp-reset

# nft list ruleset
# Warning: table ip filter is managed by iptables-nft, do not touch!
table ip filter {
	chain INPUT {
		type filter hook input priority filter; policy accept;
		counter packets 0 bytes 0 xt target REJECT
	}
}

Simply ignore the ruleset in the case that it appears to have been
generated by iptables-nft.

Signed-off-by: Kerin Millar <kfm@plushkava.net>
Signed-off-by: Sam James <sam@gentoo.org>