dev-libs/openssl: generate fipsmodule.cnf in pkg_preinst
This file contains an hash of the fips.so module, which may change after
src_install due to automatic stripping by the package manager.

README-FIPS.md says this file should be generated on each machine
separately anyway.

This will fail when cross-compiling since we call openssl from ${D}, but
I don't see a better way to handle it.

Only the primary ABI is handled currently: by the time we get to
pkg_preinst, the tools from the secondary ABIs have been clobbered.

Closes: https://bugs.gentoo.org/900625
Signed-off-by: Mike Gilbert <floppym@gentoo.org>