Revert "www-client/qutebrowser: install apparmor profile"
This reverts commit de3188d3b33e3826a278b3426d5376311439e3dc.

Adding this was a request but profile is apparently unmaintained
upstream and installing it does not seem like a good idea if it's
going to potentially crash the browser wrt bug #900861.

From a quick look it seems like it'd need adjustment for the
paths we uses too, but not limited to these (handling both slots
of qtwebengine and gpu-related things it does seems non-trivial,
lot of this profile feels like wishful guesswork).

Users that really want this are invited to install/modify it
manually for now.

Bug: https://bugs.gentoo.org/900861
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>