dev-libs/libwacom: fix python-any-r1 usage
When upgrading an old system, there's a period where Python may not
be usable at all, so either undeclared deps on Python or excessive
checks will cause failures.

In this case, Python is only BDEPENDed on for tests, and python_check_deps
tries to reflect that, but python-any-r1 is still going to check for the
interpreter. Switch back to the way the ebuild used to do things by guarding
the python-any-r1_pkg_setup call, like in the Python Guide.

This stops us checking for Python even when we're not going to use it.

Signed-off-by: Sam James <sam@gentoo.org>