www-client/qutebrowser: warn in 2.5.4 if downgrade from Qt6
Should have very few users actually using Qt6 (USE is masked,
and ebuilds are unkeyworded), but given removed the _pre ebuild
it's possible it'll cause an unexpected downgrade and it wouldn't
hurt to warn (albeit should be fairly noticeable given the deps
pulled are non-trivial).

Switching will be less of a worry in qutebrowser-3.0.0 given it
can detect a mismatch on startup, but the old 2.5.4 is oblivious.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>