pypi.eclass: Introduce PYPI_PN to override the default project name
Introduce a convenience PYPI_PN variable that can be used to override
the default project name.  This is meant to be used primarily when
upstream project name does not conform to Gentoo package naming rules,
e.g. contains dots or uppercase letters.

For example, instead of:

    SRC_URI="$(pypi_sdist_url --no-normalize "${PN/-/.}")"
    S=${WORKDIR}/${P/-/.}

one can now specify:

    PYPI_NO_NORMALIZE=1
    PYPI_PN=${PN/-/.}

For PEP 625-conformant packages, instead of:

    SRC_URI="$(pypi_sdist_url "${PN/-/.}")"

one can use:

    PYPI_PN=${PN/-/.}

There's not much gain space-wise but it avoids having to specify
the name twice.  This can particularly be helpful for package names
using PascalCase.

Signed-off-by: Michał Górny <mgorny@gentoo.org>