acct-user/deluge: add /var/lib/deluge as ACCT_USER_HOME
Sets homedir for the 'deluge' user to '/var/lib/deluge'
which is the default working directory for both deluged and
deluge-web when started with openrc. This change makes the openrc
scripts work out-of-the-box after emerging deluge.

Previously users were required to manually create and set
permissions for this directory and/or edit configs for deluged and
deluge-web for them to be startable via openrc. Since the deluge
user also writes logs to this directory by default, this solution
is not always apparent to users.

Closes: https://bugs.gentoo.org/828507
Signed-off-by: kolhell <kolspambox@gmail.com>
Closes: https://github.com/gentoo/gentoo/pull/30272
Signed-off-by: Joonas Niilola <juippis@gentoo.org>