app-emulation/libvirt: Fix a build failure with LLD 16
With the newest LLD, the libvirt's detection of --version-script=
linker argument works no more. Backport the upstream patch that
fixes the issue.

Since the patch is merged and is going to be contained in the
upcoming release, there's no point in adding it to the list of
patched in the live ebuild. Append it only into the older
versions then.

Also, I had to rebase the patch for older version of libvirt,
because the patched file (meson.build) looked a bit different and
the upstream patch did not apply cleanly.

Closes: https://bugs.gentoo.org/902211
Signed-off-by: Michal Privoznik <michal.privoznik@gmail.com>