sys-auth/polkit: keepdir /etc/polkit-1/rules.d
Calling keepdir /usr/share/polkit-1/rules.d is pointless because the
package already installs 50-default.rules there.

We want to keepdir /etc/polkit-1/rules.d so that we can ensure the
proper permissions are applied.

Skip setting the mode and owner in src_install if we are not running as
root. This resolves a failure on prefix systems.

Keep the chmod/chown in pkg_postinst to handle upgrades and ensure the
permissions get set if packages are installed out-of-order.

Move both calls behind the 'daemon' USE flag.

Closes: https://github.com/gentoo/gentoo/pull/30298
Signed-off-by: Mike Gilbert <floppym@gentoo.org>