net-analyzer/suricata: make rule-file and update dirs setuid suricata
So that it is possible to run suricata-update as root (which according
to upstream documentation is still very much allowed) but have suricata
itself drop its privileges, without having to manually change the
ownership of downloaded files. In the long run it would be nice for
suricata-update to drop privileges as well - but that's something
for upstream to take care of, and setuid suricata on the relevant
directories appears to work fine.

Closes: https://bugs.gentoo.org/900627
Signed-off-by: Marek Szuba <marecki@gentoo.org>