dev-python/cython: workaround parallel install race for importing pythran
An empty directory in site-packages is a valid Python module, unfortunately, because
of namespaces. If installing packages in parallel, the pythran module might "exist"
but be empty (hence no __version__ attribute). Catch AttributeError to avoid a narrow
race.

This might still be a Portage issue if there's a substantial delay between creating
the directory and installing the remaining files though.

Thanks to Eli Schwartz for helping out.

Bug: https://bugs.gentoo.org/902713
Signed-off-by: Sam James <sam@gentoo.org>