dev-ruby/shoulda-matchers: add 5.3.0
I spent several hours today trying to get tests running for this,
buried deep in layers-upon-layers of test dependencies, only to find
that its test suite is so dependent on being able to execute bundler
that none of it is really usable within Portage anyway.

In CI, they actively test with newer rubies and tests pass, and reverse
dependencies in Gentoo work with them enabled, so that's good enough
in this case, having tried and failed to get the tests running myself.

(I still have the branch if anyone wants to carry on with it, but
it involved packaging several NIH test deps.)

Signed-off-by: Sam James <sam@gentoo.org>