dev-util/maturin: re-enable some tests
Turns out was wrong about cli_tests, it doesn't ignore COLUMNS=80
and rather needs COLUMNS=100 to work right. Unsure what cause it to
use 100 and then do 80 if COLUMNS is unset, perhaps the terminal that
was used at the time of generation or portage internals are confusing
its defaults (not seeing 100 set anywhere in maturin's source/ci?).
Also value may not necessarily be right, 100 is just observed as the
longest lines in tests/cmd/*.stdout.

As for cargo lock, this been skipped since before I picked up
maintenance and never quite knew why it fails "sometimes" (may be
specific to rust version given I sometime ran into it only while
stabilizing using stable rust). Seems to pass right now but may
skip again if it comes back as it's not a big deal.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>