net-analyzer/nessus-agent-bin: drop fetch restriction
Tenable now offer a downloads API which can be used to retrieve most
packages without having to explicitly agree to licensing terms
beforehand. On the other hand, said terms talk about the customer being
permitted to make *one* copy of the software so let's play it safe and
keep the mirror restriction.

Signed-off-by: Marek Szuba <marecki@gentoo.org>