dev-util/ltrace: add 0.7.91_pre20221216
First attempt to move away from Debian tarballs and towards the somewhat
revived upstream. Unfortunately upstream has yet to release any versions
newer than 0.7.3 so we have to use a git snapshot for that. The suffix
_pre, with the date being that of the chosen commit, has been added
because the version number 0.7.91 got set in the repository almost a
decade ago and it is unknown at this point whether upstream intends to
increment it before the next release or not, assuming they ever make
one.

Tests are still restricted because even though they do not fail as badly
as 0.7.3 ones, all the reasons for having restricted them in the first
place remain valid.

Signed-off-by: Marek Szuba <marecki@gentoo.org>