dev-ruby/websocket-driver: EAPI 8, enable ruby32
Crank up to ruby32 by adding an upstream patch (now committed,
wasn't in the past when the each_ruby_test workaround was added
to 0.7.5). We could've just repeated the skip/quirk as was done
for ruby31, but no point if there's a fix we can pick up that's
been committed now.

Signed-off-by: Sam James <sam@gentoo.org>