net-misc/icaclient: enable usb redirect support
We need to copy the content of the usb/ subfolder into ICAROOT and
enable the USB support in module.ini. There are udev rules and a daemon
but none of that seems needed.
Users need to be in 'usb' group, mention that in postinst.

Since we have an increasing number of changes to module.ini and the user
might in fact also want to change stuff (like i.e. FIDO2), we move the
file to /etc. Same goes for usb.conf in which users can customize which
USB devices can be redirected.

Closes: https://bugs.gentoo.org/890539
Signed-off-by: Henning Schild <henning@hennsch.de>