app-emulation/virtualbox-extpack-oracle: allow old version to be installed with newer vbox
Windows 10 and 11 guests hang on >=7.0.6 additions when 3D acceleration
is disabled. The only workaround found besides enabling 3D acceleration is to
downgrade additions to 6.1.42 in the guest. This change allows the older
extpack (additions) to be installed with newer VirtualBox.

With this change, the Extension Manager will show a conflict in the but
otherwise causes no issues.

Enabling 3D acceleration is not an acceptable workaround to many due to
multiple long-standing graphical bugs.

See tickets and the forum thread below:

Freezes:
https://www.virtualbox.org/ticket/21291
https://www.virtualbox.org/ticket/21431
https://www.virtualbox.org/ticket/21448
https://www.virtualbox.org/ticket/21492

Artifacts / 3D Accel:
https://www.virtualbox.org/ticket/21136
https://www.virtualbox.org/ticket/21233
https://www.virtualbox.org/ticket/21433
https://www.virtualbox.org/ticket/21515

https://forums.virtualbox.org/viewtopic.php?f=7&t=108440

Signed-off-by: Andrew Udvare <audvare@gmail.com>
Closes: https://github.com/gentoo/gentoo/pull/30439
Signed-off-by: Viorel Munteanu <ceamac@gentoo.org>