sgml-catalog-r1.eclass: Strip ROOT when generating catalog
When cross compiling by setting the ROOT variable, the eclass was
writing the full EROOT path into the catalog file. This results in an
invalid path at runtime.

i.e.,
    $ cat /build/amd64-host/etc/sgml/catalog
    CATALOG "/build/amd64-host/etc/sgml/sgml-docbook.cat"
    CATALOG "/build/amd64-host/etc/sgml/sgml-ent.cat"
    CATALOG "/build/amd64-host/etc/sgml/xml-docbook-4.1.2.cat"

Instead we should be stripping off the ROOT so we get a valid path:

    $ cat /build/amd64-host/etc/sgml/catalog
    CATALOG "/etc/sgml/sgml-docbook.cat"
    CATALOG "/etc/sgml/sgml-ent.cat"
    CATALOG "/etc/sgml/xml-docbook-4.1.2.cat"

We don't strip EROOT because we want to keep the prefix if it's
present.

Closes: https://bugs.gentoo.org/903747
Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Closes: https://github.com/gentoo/gentoo/pull/30462
Signed-off-by: Michał Górny <mgorny@gentoo.org>