dev-ruby/mail: fix 2.7.x compat w/ newer psych (+ enable ruby31, enable ruby32)
mail-2.7.x is still needed for Rails. It turns out that this got broken
by newer Psych so tests were failing for ruby27 too. Backport the upstream
patch (actually, use Debian's backport, done for us) which is pretty
conversative and just adds a shim for newer Psych to set the needed
attributes.

Closes: https://bugs.gentoo.org/835322
Signed-off-by: Sam James <sam@gentoo.org>