dev-lang/python: skip test_tools for PGO
Skip the 'test_tools' test when running PGO for now to avoid
hanging the build (and also kind of fork-bombing the system
with recursive cpython build attempts).

Interestingly, not seen this when running the actual regular
testsuite, but I suppose far fewer people actually run that,
so could just be a frequency thing.

Bug: https://bugs.gentoo.org/828535
Bug: https://bugs.gentoo.org/850154
Bug: https://bugs.gentoo.org/903890
Closes: https://bugs.gentoo.org/900429
Signed-off-by: Sam James <sam@gentoo.org>