eclass/distutils-r1: Add PYTHON_DEPS to DEPEND unconditionally
Python distutils packages that build C extensions are currently not
declaring PYTHON_DEPS as part of their DEPEND declaration. This results
in build errors when cross compiling using ROOT=.

i.e.,
   In file included from src/setproctitle.c:14:
   In file included from src/spt.h:15:
   src/spt_python.h:15:10: fatal error: 'Python.h' file not found
   #include <Python.h>
            ^~~~~~~~~~
   1 error generated.

Since the distutils-r1 eclass currently sets the RDEPEND and BDEPEND
variables it makes sense to have the eclass also set the DEPEND
variable. We unconditionally add it to keep the API simple. If in the
future this is problematic, we can maybe add a DISTUTILS_PURE_PYTHON
eclass variable that will omit the DEPEND.

Signed-off-by: Raul E Rangel <rrangel@chromium.org>