app-emulation/libvirt: Explicitly disable openwsman in configure
Libvirt's configure script (well, meson.build file) is written so
that it automatically enables features found on the build host,
unless explicitly disabled on the cmd line. And one of such
features is 'openwsman' [1], which isn't packaged in the portage,
but if installed from an overlay, then `quickpkg` won't work.

Explicitly disable openwsman, until there's an official package
(possibly not soon [2]).

1: https://openwsman.github.io/
2: https://bugs.gentoo.org/430312

Bug: https://bugs.gentoo.org/430312
Closes: https://bugs.gentoo.org/904082
Signed-off-by: Michal Privoznik <michal.privoznik@gmail.com>
Closes: https://github.com/gentoo/gentoo/pull/30558
Signed-off-by: Sam James <sam@gentoo.org>