dev-util/rr: drop epoll_pwait2 backport
I only chucked it in because I (wrongly) thought it was needed for the
GCC 13 workaround to cleanly apply, but it isn't. I was just confused
by the spelling commit thing.

We could keep this but as jubalh points out, there's a bunch of subsequent
commits that fixup and enhance epoll_pwait2 support, so let's just leave
it for now & drop it.

I've also asked upstream if there's a release forthcoming & such, so
let's see what happens with that.

Bug: https://github.com/rr-debugger/rr/issues/3494
Fixes: f5e0ac5711c94279073106adf04977cb07db37c7
Signed-off-by: Sam James <sam@gentoo.org>