sys-devel/gcc: backport two more patches for 13 (checking ICE + Clang miscompilation)
First fix is a checking ICE. As we're new to handling checking (see previous commits),
these are generally a sign of possible wrong-code/miscompilation, so it's nice
when checking flags it up and turns the issue into an ICE. So, don't be fooled
by the ICE part.

Second is a miscompilation of Clang which leads to a test failure.

I wouldn't normally backport so many of these, but they're all things we've
found through testing in Gentoo over the last few days, and it's especially
important given GCC 13 is imminent.

Bug: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=109473
Bug: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=109462
Signed-off-by: Sam James <sam@gentoo.org>