dev-python/PyQt6: use binding operator for qtbase
Seems to rarely be an issue but hit symbol problems when using
PyQt6-6.4.2 with Qt6.5 without a rebuild (otherwise works with
6.5 just fine).

Could happen with 6.6+ too, albeit mitigated by PyQt6 bumps happening
around same time and forcing the rebuild either way.

Going to skip revbump for now given the above and being a slow build
(plus it's masked), uncertain whether it'll even be needed for 6.6.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>