app-shells/zsh: keyword ~arm64-macos.
zsh is the default login shell on macOS. When the command startprefix
is executed, by default, Gentoo prefix attempts to install the same
shell within the prefix automatically. But it currently fails due to
missing keywords. This commit adds the missing ~arm64-macos keyword
to zsh. It's tested and shown to be functional.

Bug: https://bugs.gentoo.org/904474
Signed-off-by: Yifeng Li <tomli@tomli.me>
Closes: https://github.com/gentoo/gentoo/pull/30730
Signed-off-by: Sam James <sam@gentoo.org>