dev-qt/qtwebengine: fix qutebrowser's hinting issue (6.5 regression)
Without this, scripts that qutebrowser (or others) inject may
semi-randomly fail leading to e.g. keyword hints not working, or
users' greasemonkey scripts not being loaded.

qutebrowser has opted to not do messy mitigations (at least for now)
and a fix here makes more sense. Albeit fix was not merged upstream
yet, "hopefully" will make it in 6.5.1 if doesn't miss the window.

Given qtwebengine is not used for much in Gentoo beside qutebrowser
so far, should be worth the revbump/rebuild.

Acked-by: Jimi Huotari <chiitoo@gentoo.org>
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>