dev-ruby/activesupport: fix broken minitest dependency
The minitest restriction is only required for the activesupport tests.
The code that ships with activesupport in relation to minitest can be
used by newer versions without problems.

This fixes problems where e.g. the minitest plugin in activesupport is
picked up by a test suite using a newer minitest.

Signed-off-by: Hans de Graaff <graaff@gentoo.org>