mail-mta/opensmtpd: fix libdir path
This fixes the path from the previous commit that sets the libevent and
libssl paths explicitly to avoid configure failures on multilib systems
that use lld where the incorrect lib directory is linked.

The configure check will test the path as a sysroot and then will take
it in its entirety in the 'else' case.

Fixes: https://github.com/gentoo/gentoo/commit/a05af7c36a3e4e74879cad3c03ef4ca2e3601b82
Bug: https://bugs.gentoo.org/739876
Signed-off-by: orbea <orbea@riseup.net>