distutils-r1.eclass: ensure meson-python uses a stable build directory
When not passing an explicit build directory, meson-python configures
meson into a build directory that is rooted in a tempdir with a random
suffix. This results in unstable paths, which currently disappear after
meson-python completes (either with success or failure). It doesn't
handle incremental or resumed builds, and it's very difficult to debug
-- for example by inspecting the logfile that meson itself prints out
when the configure stage fails.

Bug: https://bugs.gentoo.org/905396
Signed-off-by: Eli Schwartz <eschwartz93@gmail.com>
Signed-off-by: Sam James <sam@gentoo.org>