*/*: Add `[profile?]` to dev-lang/ghc RDEPEND
Add `[profile?]` to every dev-lang/ghc RDEPEND to make sure ghc was
built with profiling support when USE=profile is used.

Virtually every Haskell package has a dependency on the 'base' library
that ships with dev-lang/ghc. As a result, there will be errors if
dev-lang/ghc was not built with USE=profile when a Haskell package uses
the same.

Closes: https://bugs.gentoo.org/735172
Closes: https://bugs.gentoo.org/905132
Closes: https://bugs.gentoo.org/905133
Reported-by: Filip Kszczot <filip@kszczot.pl>
Signed-off-by: hololeap <hololeap@users.noreply.github.com>