net-misc/openssh-contrib: new package, add 9.3_p1
This package will include the three big third-party patch series for
HPN/SCTP/X509 functionality in OpenSSH. Historically, these patches
have caused numerous issues for users in the OpenSSH package and they
are of questionable quality. By maintaining these patches in a
separate package, we can minimize the effect of them on the garden
path, which should be to provide our users with a minimally patched
OpenSSH experience. Furthermore, since vanilla OpenSSH package will
not require a large chunk of rebasing for these patches, we can more
easily bump OpenSSH for new releases.

Signed-off-by: David Seifert <soap@gentoo.org>
Signed-off-by: Sam James <sam@gentoo.org>