games-emulation/pcsx2: switch to bundled libfmt for now
May or may not be an easy fix for libfmt-10, but upstream currently
does not really support using system libraries unless someone else
does the fix for them (e.g. closed the rapidyaml-0.5 issue and
left it broken for months).

Given libfmt tend to be volatile and break old code often, switch
to bundled for now. May revisit this later when there's non-nightly
releases, or if (at least) gets fixed for libfmt-10.

Do a revbump given portage does not fully understand rebuilding pcsx2
as being a solution to allow bumping libfmt if the dep is gone.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>