dev-python/mkdocstrings: revise relationship to mkdocstrings-python
Upstream, both packages depend on one another. However:
 - mkdocstrings dependency on mkdocstrings-python is optional
 - mkdocstrings-python does in fact import mkdocstrings code, whereas
   mkdocstrings only references mkdocstrings-python in tests

Between the above and the fact it would be a bit redundant to have
mkdocstrings ebuilds have USE flags for every single language handler
we have in the tree, drop IUSE=python + only have mkdocstrings-python in
BDEPEND (and even there, only if tests are to be run).

Signed-off-by: Marek Szuba <marecki@gentoo.org>