media-gfx/ueberzugpp: new package, add 2.8.0
This is a more versatile+faster ueberzug, albeit with heavier
dependencies and some users may prefer the original if it does
all they need.

>=net-misc/ytfzf-2.6.0 recognizes the pp version and supports
more features with it.

Still untested with masked libfmt-10 (assume may be broken), but
can't easily try given spdlog (also a dependency) is itself broken
with it.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>