media-libs/vips: new package, add 8.14.2
Adding for media-gfx/ueberzugpp, but is either way a generally nice
library to have in the tree with some useful standalone cli tools.

Debated between the naming libvips(github+website) and vips(tarball),
but the latter makes things a bit simpler if using the release tarball
and overlays have already used media-libs/vips and so can stay
consistent (also uses vips.pc). On repology it seems to be a 1/2 mix.

Preemptively using a subslot as soname been bumped several times in
the past.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>