media-libs/vips: add xorg-proto dep for USE=pango/svg/pdf
Not really used directly but results in meson error'ing out if
missing when checking for pango, cairo (also done with USE=svg),
or poppler -- or at least if pango/cairo were built with USE=X.

Unless a new EAPI adds a way for e.g. cairo to tell revdeps to pull
xorg-proto if it was built with USE=X, then not much else to do than
depend on it either way. Ultimately it's just a header-only package
and feel it's not worth risking errors over missing it.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>