gui-libs/wlroots: rework backends system, fix building with seatd
So, the initial issue was in compilation failure if seatd USE flag was
disabled - it didn't do anything (in wlroots <=0.16.2 seatd is mandatory).
In 9999 version it is optional, though, but is required for DRM and
libinput, so I reworked backends system a bit.

I have done a bit of a testing - all the ebuilds for wlroots (even the
ones I haven't touched) were successfuly recompiled (except for
unsatisfying required USE) with any combination of the following USE
flags I added: drm, libinput, session.

Bug: https://bugs.gentoo.org/883781
Closes: https://bugs.gentoo.org/883781
Signed-off-by: Zurab Kvachadze <zurabid2016@gmail.com>