profiles: create USE=valgrind global USE flag
This always has the same meaning in packages - build in annotations to help
with e.g. custom memory allocators to reduce noise and improve Valgrind's accuracy.

All invalid uses of this were already fixed (cases where it was used to control
running the testsuite under Valgrind which we don't want to do, it's too flaky
under sandbox & not reliable with diff arches.)

Signed-off-by: Sam James <sam@gentoo.org>