dev-python/cryptography: Backport pyo3-0.18 port, for pypy3.10
Backport a port to pyo3-0.18.3 that is necessary to fix compatibility
with PyPy3.10.  This is because pyo3 is a quality package that relies
on implementation details of PyPy, and 0.18.2 is the first version
to support PyPy3.10.  Unfortunately, the code of cryptography-40.0.2
is incompatible with pyo3-0.18, so we need to backport the whole 46k
patch porting it.

Signed-off-by: Michał Górny <mgorny@gentoo.org>