gui-libs/wlroots: fix building with seatd, revbump to wlroots-0.16.2-r1
Previous ebuild would conditionally, based on the USE flags, pull
mandatory dependencies (that is udev, hwdata, seatd), which in the
absense of such would result in a build failure. USE flags have been
removed, which does not affect users who managed to build the package
successfully, therefore keywords were not dropped.

It is possible to disable/enable certain backends including their
dependencies, so ability to control this behaviour was added in 0.16.1-r1
and 9999 versions.

Bug: https://bugs.gentoo.org/883781
Closes: https://bugs.gentoo.org/883781
Signed-off-by: Zurab Kvachadze <zurabid2016@gmail.com>
Signed-off-by: Matthew Thode <prometheanfire@gentoo.org>