profiles/targets/desktop: enable USE=vulkan by default
Ionen pointed this out again today and it made me look back at it; there's
no reason to not have vulkan on by default for desktop profiles.

In particular, Steam expects it for plenty of games to work, but at this point,
it's becoming jarring for users to _not_ have it as it's considered a fundamental part
of the modern desktop.

Signed-off-by: Sam James <sam@gentoo.org>