dev-php/pecl-uploadprogress: fix tests with PHP's --enable-debug.
A few of the tests for pecl-uploadprogress throw a fatal error if
PHP was build with USE=debug. We skip the tests for now; a better
long-term solution was proposed upstream (GH issue #17).

In addition, the extension seems to work fine with php-8.2, so I've
taken the opportunity to enable it.

Closes: https://bugs.gentoo.org/906691
Signed-off-by: Michael Orlitzky <mjo@gentoo.org>