dev-util/spirv-llvm-translator: Remove IUSE=tools
It has not worked since upstream commit ae904012 ("[CMake] Enable
LLVM_BUILD_TOOLS for out-of-tree builds") in v15.0.0.

With that commit reverted, USE=-tools saves building one 120 KiB executable
(llvm-spirv) which is a single source file.

Also unconditionally depend on dev-util/spirv-tools as it is an
automagic dependency of llvm-spirv which will always be installed for
consumers of dev-util/spirv-llvm-translator.

Signed-off-by: Matt Turner <mattst88@gentoo.org>