profiles/use.desc: create USE=modules-sign global USE flag
Similarly to gyakovlev's proposition for signing back in 2018 (with
a module-sign IUSE), linux-mod-r1.eclass will make use of this to
enable/disable signing and it would be inconvenient if consumers
had to define it.

An alternative could be to automagic enable when the kernel has
"sign by default" a bit like compression is handled -- albeit this
can sometime need more configuration and may be unexpected (i.e.
permissions for keys, if keys were moved to a different locations,
passphrases, and dist-kernels unsurprisingly don't install the
private key and would result in failure out-of-the-box).

Having a USE also makes it more obvious that support exists,
and attempting to enable will give bit of explanations if anything
is amiss.

Name-wise, debated between this and 'sign-modules' but fwiw former
sorts better with the already existing 'modules'.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>