sys-fs/z̸͚͇̾̾͌f̵̦̠͌̾̕s̴̢̪̙̒͌-̴̻͓̒̒̚k̵͓͔͊̀̈́m̵̢̪̟͋́̚o̸̦̦̟̓̈́͛d̸̡͉̼͌͆͑: (demo migration, not for merge)
Base migration for reference, albeit bit of a hack. Whether it
should be used is up to debate but I think it's not so bad and
patch is fairly trivial.

Not planning to finish/push this myself (anyone is welcome to
take these commits and test/adapt), I don't use it and have not
even tried to load the modules (albeit it built fine with
clang16+thinlto).

Use at you own risks.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>