sys-fs/zfs-kmod: further linux-mod-r1 adaptations & style tweaks
Just meant to integrate MODULES_KERNEL_MIN+MAX ended up making quick
changes to my liking while at it. Thus the separate commit (also is
barely tested changes). Probably can still use polishing.

MAX is non-fatal, but gives a very large warning that is hard to miss,
imo removing the need for a hidden override variable. Made optional
for dist-kernel through another USE so people can still get automated
rebuilds if they e.g. patched it (pondered adding a similar mechanism
builtin the eclass but believe too few ebuilds really need this and
don't want to encourage it globally unless really needed).

wrt <virtual/dist-kernel, haven't looked into the history of it but
right now can't quite see why it need to be BDEPEND, regardless of it
existing in the eclass portage will have to respect the extra bound in
RDEPEND as this isn't a || ( ). Or at least, it's not trying to upgrade
my kernel here and I get the usual "WARNING:" that it can't upgrade
(on a side-note, don't need to repeat the := given the eclass' is
still valid).

(sam: preserved ionen's original commit msg for posterity but obviously
tested it and reviewed it since. We discussed the cap situation and concluded
it's probably best for now to stick with it as-is (in this commit) given +dist-kernel-cap
handles it for most cases, and if not using dist-kernel, you may not want things
managed for you anyway.)

Bug: https://bugs.gentoo.org/814194
Bug: https://bugs.gentoo.org/865157
Signed-off-by: Ionen Wolkens <ionen@gentoo.org>
Signed-off-by: Sam James <sam@gentoo.org>