app-admin/ryzen_smu: migrate to linux-mod-r1
I don't actually use this (no hardware), but migrating myself as an
example given it's simple and cleanups the CC_IS_CLANG checks. Sorry
if it breaks anything (feel free to mask or revert if need be).

Haven't heard from slashbeast (maybe didn't notice PR), but hopefully
this is okay.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>