x11-drivers/nvidia-drivers: soften IBT warning
The situation around this is still unclear, getting conflicting
reports and (if actually fixed), it's still unclear which versions
have received backports.

Does not help that, at least initially, it was only fixed with
USE=kernel-open because (unlike nvidia last I know of) we are
using a newer gcc. And then changelog entries often have mixups
(and wrt news outlets, afaik all they did is read these).

If anyone reading this actually uses IBT, please send a report
if the blob works as intended.

Not changing 390's ebuild given that one is almost certainly broken.

Signed-off-by: Ionen Wolkens <ionen@gentoo.org>